v0.3.1: Restructure Babel configuration
The recently introduced `eslint-import-resolver-babel-module` package
processes Babel configuration files all the way up the dependency tree.
When linting an application using `bitumen`'s rules, the resolver would
fail when attempting to process `bitumen`'s own Babel configuration,
since it referenced an internal-use file that was not present in the
`bitumen` distributable.

To avoid such errors in downstream consumers, this change restructures
the exported Babel configuration so that the file it imports is present
in the distributable and directly accessible from outside the package.